--- case_study/src/main/java/com/example/case_study/service/IEmployeeTypeService.java
@@ -1,9 +1,0 @@
-package com.example.case_study.service;
-
-import com.example.case_study.model.EmployeeType;
-
-import java.util.List;
-
-public interface IEmployeeTypeService {
-    List<EmployeeType> getAll();
-}
