+++ castudy_module_4/src/main/java/com/example/castudy_module_4/repository/IEmployeeRepository.java
@@ -1,6 +1,6 @@
-package com.example.case_study.repository;
+package com.example.castudy_module_4.repository;
 
-import com.example.case_study.model.Employee;
+import com.example.castudy_module_4.model.employee.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IEmployeeRepository extends JpaRepository<Employee,Integer> {
