+++ castudy_module_4/src/main/java/com/example/castudy_module_4/repository/IEmployeeTypeRepository.java
@@ -1,6 +1,6 @@
-package com.example.case_study.repository;
+package com.example.castudy_module_4.repository;
 
-import com.example.case_study.model.EmployeeType;
+import com.example.castudy_module_4.model.employee.EmployeeType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IEmployeeTypeRepository extends JpaRepository<EmployeeType,Integer> {
