+++ castudy_module_4/src/main/java/com/example/castudy_module_4/service/IEmployeeService.java
@@ -1,0 +1,18 @@
+package com.example.castudy_module_4.service;
+
+import com.example.castudy_module_4.dto.employeeDTO.EmployeeDTO;
+import com.example.castudy_module_4.model.employee.Employee;
+
+import java.util.List;
+
+public interface IEmployeeService {
+    List<Employee> getAll();
+
+    void save(Employee employee);
+
+    void delete(int id);
+
+    Employee findById(int id);
+
+    void update(Employee employee);
+}
