+++ castudy_module_4/src/main/java/com/example/castudy_module_4/service/IEmployeeTypeService.java
@@ -1,0 +1,9 @@
+package com.example.castudy_module_4.service;
+
+import com.example.castudy_module_4.model.employee.EmployeeType;
+
+import java.util.List;
+
+public interface IEmployeeTypeService {
+    List<EmployeeType> getAll();
+}
