+++ castudy_module_4/src/main/java/com/example/castudy_module_4/service/employeeImpl/EmployeeService.java
@@ -1,9 +1,9 @@
-package com.example.case_study.service.impl;
+package com.example.castudy_module_4.service.employeeImpl;
 
-import com.example.case_study.model.Employee;
-import com.example.case_study.repository.IEmployeeRepository;
-import com.example.case_study.repository.IEmployeeTypeRepository;
-import com.example.case_study.service.IEmployeeService;
+import com.example.castudy_module_4.model.employee.Employee;
+import com.example.castudy_module_4.repository.IEmployeeRepository;
+import com.example.castudy_module_4.repository.IEmployeeTypeRepository;
+import com.example.castudy_module_4.service.IEmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,5 +27,15 @@ public class EmployeeService implements IEmployeeService {
     @Override
     public void delete(int id) {
         employeeRepository.deleteById(id);
+    }
+
+    @Override
+    public Employee findById(int id) {
+        return employeeRepository.findById(id).get();
+    }
+
+    @Override
+    public void update(Employee employee) {
+        employeeRepository.save(employee);
     }
 }
