+++ castudy_module_4/src/main/resources/application.properties
@@ -1,5 +1,5 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/case_study_4?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false
+spring.datasource.url=jdbc:mysql://localhost:3306/vegetable_management?createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true&useSSL=false
 spring.datasource.username=root
-spring.datasource.password=anhdao1911
-spring.jpa.hibernate.ddl-auto=create
+spring.datasource.password=01282637935Aa
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
