+++ castudy_module_4/src/main/resources/templates/employees/create.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Create product</title>
+</head>
+<body>
+<form th:object="${employeeDTO}" method="post" action="/employee/create">
+    <div class="input-group">
+        <label for="fullName">Full name : </label>
+        <input type="text"
+               class="form-control" name="fullName" id="fullName" aria-describedby="helpId" placeholder=""
+               th:field="*{fullName}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{fullName}"></span>
+    </div>
+    <div class="input-group">
+        <label for="email">email : </label>
+        <input type="text"
+               class="form-control" name="email" id="email" aria-describedby="helpId" placeholder=""
+               th:field="*{email}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{email}"></span>
+    </div>
+    <div class="input-group">
+        <label for="gender">Gender : </label>
+        <input type="text"
+               class="form-control" name="gender" id="gender" aria-describedby="helpId" placeholder=""
+               th:field="*{gender}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{gender}"></span>
+    </div>
+    <div class="input-group">
+        <label for="address">Address : </label>
+        <input type="text"
+               class="form-control" name="address" id="address" aria-describedby="helpId" placeholder=""
+               th:field="*{address}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{address}"></span>
+    </div>
+    </div>
+    <div class="input-group">
+        <label for="phoneNumber">Phone Number : </label>
+        <input type="text"
+               class="form-control" name="phoneNumber" id="phoneNumber" aria-describedby="helpId" placeholder=""
+               th:field="*{phoneNumber}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{phoneNumber}"></span>
+    </div>
+    <select>
+        <option th:each="employee,inerStat :${employeeType}" th:value="${employee.getId()}">
+            <p th:text="${employee.getName()}"></p>
+        </option>
+    </select>
+    <button type="submit">Save</button>
+</form>
+
+
+</body>
+</html>
