+++ castudy_module_4/src/main/resources/templates/employees/update.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Update Employee</title>
+</head>
+<body>
+<form action="/employee/update" method="post" th:object="${employeeDTO}">
+    <div class="input-group">
+        <label for="fullName">Full name :</label>
+        <input type="text"
+               class="form-control" name="fullName" id="fullName" aria-describedby="helpId" placeholder="" th:field="*{fullName}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{fullName}"></span>
+    </div>
+    <div class="input-group">
+        <label for="email">Email :</label>
+        <input type="text"
+               class="form-control" name="email" id="email" aria-describedby="helpId" placeholder="" th:field="*{email}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{email}"></span>
+    </div>
+    <div class="input-group">
+        <label for="gender">Gender :</label>
+        <input type="text"
+               class="form-control" name="gender" id="gender" aria-describedby="helpId" placeholder="" th:field="*{gender}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{gender}"></span>
+    </div>
+    <div class="input-group">
+        <label for="address">Address :</label>
+        <input type="text"
+               class="form-control" name="address" id="address" aria-describedby="helpId" placeholder="" th:field="*{address}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{address}"></span>
+    </div>
+    <div class="input-group">
+        <label for="phoneNumber">Phone number :</label>
+        <input type="text"
+               class="form-control" name="phoneNumber" id="phoneNumber" aria-describedby="helpId" placeholder="" th:field="*{phoneNumber}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{phoneNumber}"></span>
+    </div>
+    <div class="input-group">
+        <label for="password">Password</label>
+        <input type="text"
+               class="form-control" name="password" id="password" aria-describedby="helpId" placeholder="" th:field="*{password}">
+        <span th:if="${#fields.hasErrors()}" th:errors="*{password}"></span>
+    </div>
+    <button type="submit">Update</button>
+</form>
+</body>
+</html>
