--- castudy_module_4/src/main/resources/templates/list.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<h1 th:text="${msg}"></h1>
-</body>
-</html>
